Moved Necker reverse functions into a new header.
Also fixed issues with missing core namespace
