+++ arrows/core/CMakeLists.txt
@@ -35,6 +35,7 @@ set( plugin_core_headers
   match_features_homography.h
   match_tracks.h
   mesh_operations.h
+  necker_reverse.h
   read_object_track_set_kw18.h
   read_track_descriptor_set_csv.h
   render_mesh_depth_map.h
@@ -111,6 +112,7 @@ set( plugin_core_sources
   match_features_fundamental_matrix.cxx
   match_features_homography.cxx
   match_tracks.cxx
+  necker_reverse.cxx
   mesh_operations.cxx
   read_object_track_set_kw18.cxx
   read_track_descriptor_set_csv.cxx
